requirements.txt
django-toolbelt
Django
dj-database-url
dj-static
gunicorn
psycopg2
static
django-tastypie
django-registration
django-registration-defaults
